# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API by Housni BENABID",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/main/promotions": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "List blogposts",
                "description": "List blogposts endpoint",
                "operationId": "promotions.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}